FIRST_BLOCK_RIG
First block rig is officially done. Problem areas that need to be worked on are: IK/FK Switch does not let you blend without major deformation and Hands do not orient properly and are not implemented. Other then that, we should be good to go until we see new problems. This block rig works for the Old lady and Older Girl