Commits from Wednesday and Today
Commiting a Block rig for the wolf model. Also changing "Block_Rig_Final.ma" to accurately reflect that it is, in fact, the final block rig. Added hand joints and fixed the issues with switches and hips. Still need to finalize the hand controls and set up the wolf rig, but here is the work I have today.